+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Generated migrations can be rollbacked
+- clients pkey constraint do not crash on admin create
 
 ## [2.1.2] - 2022-05-02